chore: 드래그 아이템 컴포넌트 분리
